Finalise Testimonials Home (PR #8 Changes): Integrated testimonials with Swiper component and separated testimonials data into JSON
